requirements.txt
pytest
pytest_bdd
selenium
jsonpickle
pony
pymysql
lxml-4.6.2-cp37-cp37m-win_amd64.whl
pytest-allure-adapter